backend/adminuser/migrations/serializers.py
from rest_framework import serializers
from .models import Admin

class AdminSerializer(serializers.ModelSerializer):
    class Meta:
        model = Admin
        fields = '__all__'  # Include all fields
        read_only_fields = ('joining_date',)  # This field is auto-added

    def validate_phone(self, value):
        if len(value) != 10 or not value.isdigit():
            raise serializers.ValidationError("Phone number must be 10 digits long.")
        return value
